refactor: 프로젝트 구조 변경 및 공통응답 포맷 변경
