--- src/main/java/com/example/buserve/src/chargingmethod/AmountDto.java
@@ -1,10 +1,0 @@
-package com.example.buserve.src.chargingmethod;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class AmountDto {
-    private Integer amount;
-}
