--- src/main/java/com/example/buserve/src/configure/ApiResponse.java
@@ -1,36 +1,0 @@
-package com.example.buserve.src.configure;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-import static com.example.buserve.src.configure.ApiResponseStatus.SUCCESS;
-
-@Getter
-@AllArgsConstructor
-@JsonPropertyOrder({"isSuccess", "code", "message", "result"})
-public class ApiResponse<T> {
-    @JsonProperty("isSuccess")
-    private final Boolean isSuccess;
-    private final String message;
-    private final int code;
-    @JsonInclude(JsonInclude.Include.NON_NULL)
-    private T result;
-
-    // 요청에 성공한 경우
-    public ApiResponse(T result) {
-        this.isSuccess = SUCCESS.isSuccess();
-        this.message = SUCCESS.getMessage();
-        this.code = SUCCESS.getCode();
-        this.result = result;
-    }
-
-    // 요청에 실패한 경우
-    public ApiResponse(ApiResponseStatus status) {
-        this.isSuccess = status.isSuccess();
-        this.message = status.getMessage();
-        this.code = status.getCode();
-    }
-}
