+++ src/main/java/com/example/buserve/src/pay/dto/AmountDto.java
@@ -1,0 +1,12 @@
+package com.example.buserve.src.pay.dto;
+
+import lombok.*;
+
+@Getter
+@Setter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class AmountDto {
+    private Integer amount;
+}
